Merge upstream changes (#2)
* Allows username validation against IAM groups

This change gives the option to validate the remote username against
the IAM groups containing the user invoking the lambda function. This
is an optional feature which is used in conjunction with kmsauth.

For example, if there were two groups of users, you could put your
admins in the ssh-admin IAM group to allow them to generate certificates
with a remote_username of 'admin'. Users with fewer permissions could be
in the ssh-user group to allow them to generate certificates for the 'user'
account.

The group name is configurable, however they must all be in a consistent
format, and must all contain the relevant remote_username once.

* Compressed CA private key support

* Fixing https://github.com/Netflix/bless/issues/72 thanks @Immortalin and @tuxinaut .

* Add support for loading ED25519 public keys

* Add certificate builder and test ED25519 signed by RSA

* Allowing BLESS lambda to accept ed25519 keys, completing https://gith… (#74)

* Allowing BLESS lambda to accept ed25519 keys, completing https://github.com/Netflix/bless/pull/71 .  Thanks @jnewbigin .

* Moving BLESS to python 3.6. (#75)

* Moving BLESS to python 3.6.
You just need to rebuild, publish, and switch your lambda runtime from 2.7 to 3.6.

* Moving TravisCI to Python3.6 as well.

* bless_client.py: fix argv unpacking when using a kmsauth token (#63)

* Add the FileSync flag to the zip command (#76)

* Make lambda_configs dir optional for publish make target (#69)

* Adding a blacklisted remote_usernames option.  This would prevent particular SSH Authorized Principals from being included in a BLESS certificate.

* Refactored BLESS to cache KMS decrypt results for the ca private key password.

* Bumping to Release v.0.3.0

Features include:
Python 3.6 Lambda support
Caching of the KMS decrypted CA Private Key Password.
Compressed CA Private Key support, allowing RSA 4096 keys to be set in the Lambda Environment.
Issue certificates for ED25519 public keys (RSA CA).
New option to validate the remote username against the IAM groups of the calling user.
Updated dependencies.